+++ changelog.md
@@ -90,6 +90,7 @@ Codion Change Log
 - FilterTable bug fixed, sorting, column reordering and resizing key events were enabled even though the functionality had been disabled.
 - TableConditionPanel.focusEvent() added.
 - FilterColumnConditionPanel now initialized lazily.
+- FilterTableConditionPanel.focusGainedEvent() moved to ColumnConditionPanel, related refactoring.
 ### is.codion.framework.domain
 - DefaultDomain renamed DomainModel.
 - Entity.valuesEqual() instance methods renamed equalValues(), static ones removed.